Fix Qt meta-object queuing issue with PurchaseError enum
Change purchaseFailed signal to use int parameter instead of PurchaseError
enum to resolve "Cannot queue arguments of type 'PurchaseError'" error.
Qt has issues with queued signal connections for custom enum types across
threads, but int parameters work perfectly.

Changes:
• Update purchaseFailed signal signature:
  - Old: purchaseFailed(PurchaseError error, int platformCode, QString message)
  - New: purchaseFailed(int error, int platformCode, QString message)
• Cast PurchaseError to int when emitting signals across all backends:
  - Android: static_cast<int>(error)
  - iOS: static_cast<int>(error)
  - Windows: static_cast<int>(error)
• Update AbstractStoreBackend constructor lambda to handle int parameter

QML usage remains unchanged:
- Can still compare: if (error === Store.UserCanceled)
- Q_ENUM registration provides QML enum access
- Type-safe comparisons work as before

Benefits:
✅ Thread-safe queued signal connections
✅ No meta-type registration required
✅ Backward compatible QML enum usage
✅ Maintains semantic error handling

+++ src/android/googleplaystorebackend.cpp
@@ -178,7 +178,7 @@ void GooglePlayStoreBackend::consumePurchase(AbstractTransaction * transaction)
 
     PurchaseError error = mapBillingResponseToPurchaseError(billingResponseCode);
     QString message = getBillingResponseMessage(billingResponseCode);
-    emit backend->purchaseFailed(error, billingResponseCode, message);
+    emit backend->purchaseFailed(static_cast<int>(error), billingResponseCode, message);
 }
 
 /*static*/ void GooglePlayStoreBackend::purchaseConsumed(JNIEnv *env, jobject object, jstring message)
